+++ tests/testthat/test-vroom.R
@@ -566,3 +566,7 @@ test_that("vroom can handle NUL characters in strings", {
     equals = tibble::tibble(abc = "ab", def = "def")
   ))))
 })
+
+test_that("n_max is respected in all cases", {
+  expect_equal(dim(vroom("x\ty\tz\n1\t2\t3\n4\t5\t6\n", n_max = 1)), c(1, 3))
+})